+++ src/withRenew.js
@@ -44,20 +44,23 @@ const withRenew = WrappedComponent => class WithRenewComponent extends React.Com
   );
 
   static propTypes = {
+    loans: PropTypes.object,
     mutator: PropTypes.shape({
-      renew: PropTypes.shape({
-        POST: PropTypes.func.isRequired,
-      }),
       loanPolicies: PropTypes.shape({
         GET: PropTypes.func.isRequired,
         reset: PropTypes.func.isRequired,
+      }),
+      renew: PropTypes.shape({
+        POST: PropTypes.func.isRequired,
       }),
       requests: PropTypes.shape({
         GET: PropTypes.func.isRequired,
         reset: PropTypes.func.isRequired,
       }),
     }),
-    loans: PropTypes.object,
+    stripes: PropTypes.shape({
+      connect: PropTypes.func.isRequired,
+    }),
   };
 
   static defaultProps = {
@@ -232,6 +235,14 @@ const withRenew = WrappedComponent => class WithRenewComponent extends React.Com
 
   hideBulkRenewalDialog = () => this.setState({ bulkRenewalDialogOpen: false });
 
+  /**
+   * retrieve count of open requests against this patron's open loans and
+   * store a map of itemId => open-request count in state.
+   *
+   * It sure would be nice if there were a more efficient way to construct
+   * this query than joining the entire list of item-ids, but there ain't.
+   * See CHAL-30 for details of the pain this causes.
+   */
   getOpenRequestsCount = () => {
     const {
       stripes,
@@ -249,8 +260,8 @@ const withRenew = WrappedComponent => class WithRenewComponent extends React.Com
       return;
     }
 
-    const q = loans.map(loan => `itemId==${loan.itemId}`).join(' or ');
-    const query = `(${q}) and status==("Open - Awaiting pickup" or "Open - Not yet filled") sortby requestDate desc`;
+    const q = loans.map(loan => loan.itemId).join(' or ');
+    const query = `(itemId==(${q})) and status==("Open - Awaiting pickup" or "Open - Not yet filled") sortby requestDate desc`;
 
     reset();
 
@@ -268,6 +279,10 @@ const withRenew = WrappedComponent => class WithRenewComponent extends React.Com
   };
 
 
+  /**
+   * retrieve loan policies related to current loans and store a map of
+   * loan-policy.id => loan-policy.name in state.
+   */
   fetchLoanPolicyNames = () => {
     // get a list of unique policy IDs to retrieve. multiple loans may share
     // the same policy; we only need to retrieve that policy once.