+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Retrieve up to 1000 loans instead of 100. LIBRARIANS LOVE BOOKS! Refs CHAL-29.
 * Dedupe loan policy list before retrieving it. Refs CHAL-30.
+* Slightly more efficient request retrieval. Refs CHAL-30.
 
 ## [2.24.1](https://github.com/folio-org/ui-users/tree/v2.24.1) (2019-07-26)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.24.0...v2.24.1)